Moved all plant-related events to api package
