+++ api/src/main/java/eu/germanrp/addon/api/events/plant/PlantPacketReceiveEvent.java
@@ -1,4 +1,4 @@
-package eu.germanrp.addon.core.common.events.plant;
+package eu.germanrp.addon.api.events.plant;
 
 import eu.germanrp.addon.api.network.PlantPacket;
 import lombok.Data;
