+++ core/src/main/java/eu/germanrp/addon/core/listener/EventRegistrationListener.java
@@ -9,9 +9,9 @@ import eu.germanrp.addon.core.Utils;
 import eu.germanrp.addon.core.common.events.GermanRPAddonTickEvent;
 import eu.germanrp.addon.core.common.events.GraffitiUpdateEvent;
 import eu.germanrp.addon.core.common.events.LegacyGermanRPUtilsPayloadEvent;
-import eu.germanrp.addon.core.common.events.plant.PlantCreateEvent;
-import eu.germanrp.addon.core.common.events.plant.PlantDestroyEvent;
-import eu.germanrp.addon.core.common.events.plant.PlantPacketReceiveEvent;
+import eu.germanrp.addon.api.events.plant.PlantCreateEvent;
+import eu.germanrp.addon.api.events.plant.PlantDestroyEvent;
+import eu.germanrp.addon.api.events.plant.PlantPacketReceiveEvent;
 import lombok.val;
 import net.labymod.api.event.Subscribe;
 import net.labymod.api.event.client.chat.ChatReceiveEvent;
