+++ core/src/main/java/eu/germanrp/addon/core/widget/PlantHudWidget.java
@@ -6,9 +6,9 @@ import eu.germanrp.addon.api.events.plant.PlantReadyToHarvestEvent;
 import eu.germanrp.addon.api.models.Plant;
 import eu.germanrp.addon.api.models.PlantFactory;
 import eu.germanrp.addon.api.models.PlantType;
-import eu.germanrp.addon.core.common.events.plant.PlantCreateEvent;
-import eu.germanrp.addon.core.common.events.plant.PlantDestroyEvent;
-import eu.germanrp.addon.core.common.events.plant.PlantPacketReceiveEvent;
+import eu.germanrp.addon.api.events.plant.PlantCreateEvent;
+import eu.germanrp.addon.api.events.plant.PlantDestroyEvent;
+import eu.germanrp.addon.api.events.plant.PlantPacketReceiveEvent;
 import eu.germanrp.addon.core.executor.PlaySoundExecutor;
 import lombok.val;
 import net.labymod.api.client.component.Component;
